pixela-java-client-63 Refactoring Codec classes - move BasicResponse from impl project to api project because the type is well known.
